requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
aiosonic==0.9.7
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
croniter==0.3.37
graia-application-mirai==0.18.3
graia-broadcast==0.8.7
graia-saya==0.0.8
h2==4.0.0
hiredis==1.1.0
hpack==4.0.0
hyperframe==6.0.0
idna==3.1
iterwrapper==0.1.4
linnian-apps-chatter==0.1.8
loguru==0.5.3
multidict==5.1.0
natsort==7.1.1
Pillow==8.1.2
Pillow-PIL==0.1.dev0
pydantic==1.7.1
python-dateutil==2.8.1
regex==2020.11.13
six==1.15.0
typing-extensions==3.7.4.3
ujson==4.0.2
yarl==1.6.3
